requirements.txt
confluent-kafka==2.3.0
cryptography==42.0.1
httpx==0.26.0
prefect==2.14.17
pydash==7.0.7
pymongo==4.6.1
python-dotenv==1.0.1
